Moved tests for transform to the right project.
Fixed bug in AWeightedTrigometricDistance setting the alpha in a wrong
way.
